refactor(165): url 구분 처리
1. app, sdk 구분되도록 수정
- user (결제 고객, app)
- merchant (상점 고객, sdk)

2. GlobalExceptionHandler 내 logger 추가

3. ddl-auto : none 설정
