+++ api/src/main/kotlin/com/inner/circle/api/exception/GlobalExceptionHandler.kt
@@ -5,6 +5,7 @@ import com.inner.circle.api.structure.dto.PaymentError
 import com.inner.circle.api.structure.dto.PaymentResponse
 import com.inner.circle.exception.AppException
 import javax.security.sasl.AuthenticationException
+import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ExceptionHandler
@@ -12,12 +13,18 @@ import org.springframework.web.bind.annotation.RestControllerAdvice
 
 @RestControllerAdvice
 class GlobalExceptionHandler {
+    companion object {
+        private val logger = LoggerFactory.getLogger(GlobalExceptionHandler::class.java)
+    }
+
     @ExceptionHandler(AppException::class)
-    fun handleAppException(exception: AppException): ResponseEntity<AppExceptionResponse> =
-        ResponseEntity(
+    fun handleAppException(exception: AppException): ResponseEntity<AppExceptionResponse> {
+        logger.info("AppException (type = {})", exception::class.simpleName, exception)
+        return ResponseEntity(
             AppExceptionResponse.of(exception),
             HttpStatus.valueOf(exception.status.code)
         )
+    }
 
     @ExceptionHandler(AuthenticationException::class)
     fun handleAuthorizationException(
@@ -26,6 +33,7 @@ class GlobalExceptionHandler {
 
     @ExceptionHandler(Exception::class)
     fun handleException(ex: Exception): ResponseEntity<PaymentResponse<Nothing>> {
+        logger.error("An unexpected error occurred", ex)
         val status = HttpStatus.INTERNAL_SERVER_ERROR
         val errorResponse =
             PaymentResponse.fail(
@@ -38,6 +46,7 @@ class GlobalExceptionHandler {
     fun handleIllegalArgumentException(
         ex: IllegalArgumentException
     ): ResponseEntity<PaymentResponse<Nothing>> {
+        logger.warn("invalid request occurred", ex)
         val status = HttpStatus.BAD_REQUEST
         val errorResponse =
             PaymentResponse.fail(
