+++ infra/src/main/resources/application-infra.yaml
@@ -59,7 +59,7 @@ spring:
       max-lifetime: 1800000 # 30분
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: none
     open-in-view: false
     properties:
       hibernate:
